refactor(export): split gltf.rs, complete from-meshes fail-closed validation
Addresses the CI failures and review feedback on the perf PR:

- Module-size ratchet: gltf.rs grew past its 4028-line budget. Move the
  instancing matrix math (+ its RTC test) into `gltf/matrix.rs` and the
  from-meshes assembler into `gltf/from_meshes.rs` (child modules reaching the
  parent's private internals via `super`). Pure code moves — GLB output stays
  byte-identical (verified against the pre-change golden across all export
  paths) and gltf.rs is back under budget.

- Complete the from-meshes fail-closed validation (review P1): reject empty or
  short `normals` (would drop every under-covered mesh — empty normals drops the
  WHOLE model) and fewer `index_counts` than declared meshes, not just
  vertex/index-count buffer overruns. The viewer always passes fully-backed,
  normal-covered arrays, so this only fires on a caller bug.

- Tests for every validation branch: short normals, empty normals, and
  index_counts overrun (in addition to the existing valid + vertex-overrun).

- Regenerate the committed wasm `ifc-lite.d.ts` for the boundary doc; the
  `exportGlbFromMeshes` TS signature is unchanged (`Uint8Array`).
