+++ rust/export/src/gltf.rs
@@ -39,6 +39,14 @@ use ifc_lite_processing::{
 use serde::Serialize;
 use serde_json::{json, Value};
 
+// Split-out submodules (kept child modules so they can reach this module's private
+// assembler internals via `super`; the code moved verbatim, so output is unchanged).
+mod from_meshes;
+mod matrix;
+
+pub use from_meshes::{export_glb_from_meshes, try_export_glb_from_meshes};
+use matrix::{affine_inverse, compose_world_meta, occurrence_node_matrix};
+
 /// Options for glTF/GLB export.
 pub struct GltfOptions {
     /// Attach `asset.extras` (counts) and per-node `extras.expressId`.
@@ -372,158 +380,6 @@ fn geom_color_key(positions: &[f32], normals: &[f32], indices: &[u32], color: [f
     h.update(&b.to_le_bytes());
     h.update(&a.to_le_bytes());
     h.digest128()
-}
-
-// ── Instancing matrix math (row-major f64 4x4) ──────────────────────────────
-//
-// An occurrence's node matrix must map the shared template's Y-up LOCAL geometry
-// to that occurrence's Y-up BAKED world position, minus the model-wide
-// `scene_center` that the root node carries:
-//
-//   N_k = T(-scene_center) · S · [ T(-rtc) · (M_k · M_ref⁻¹) · T(rtc) ] · S⁻¹ · T(template_origin_yup)
-//
-// where `M = transform · local · canonical` is the per-occurrence world placement
-// from `InstanceMeta` (Z-up, **pre-RTC**), `rtc` is the model RTC/site offset the
-// baker subtracted (Z-up), and `S` is the Z-up→Y-up basis `(x,y,z) → (x, z, -y)`.
-// The `T(-rtc)·…·T(rtc)` conjugation moves the relative transform from the pre-RTC
-// frame `M` lives in into the POST-RTC baked frame the template geometry is in —
-// without it, a rotated occurrence under a non-zero site/georef offset is
-// mis-translated by `(R_rel - I)·rtc` (kilometres at national-grid scale). Everything
-// is f64, recomputed from the f64 `InstanceMeta` (NOT the collator's f32 `rel`), so
-// the absolute-magnitude terms cancel to a small, f32-precise translation before the
-// final downcast even at national-grid coordinates.
-
-/// Z-up→Y-up basis as a row-major 4x4 (linear part only; `(x,y,z) → (x, z, -y)`).
-const S_YUP: [f64; 16] = [
-    1.0, 0.0, 0.0, 0.0, //
-    0.0, 0.0, 1.0, 0.0, //
-    0.0, -1.0, 0.0, 0.0, //
-    0.0, 0.0, 0.0, 1.0,
-];
-/// Inverse (transpose, since `S_YUP` is a proper rotation): `(x,y,z) → (x, -z, y)`.
-const S_YUP_INV: [f64; 16] = [
-    1.0, 0.0, 0.0, 0.0, //
-    0.0, 0.0, -1.0, 0.0, //
-    0.0, 1.0, 0.0, 0.0, //
-    0.0, 0.0, 0.0, 1.0,
-];
-const IDENTITY16: [f64; 16] = [
-    1.0, 0.0, 0.0, 0.0, //
-    0.0, 1.0, 0.0, 0.0, //
-    0.0, 0.0, 1.0, 0.0, //
-    0.0, 0.0, 0.0, 1.0,
-];
-
-/// Row-major 4x4 multiply `a · b`.
-fn mat4_mul(a: &[f64; 16], b: &[f64; 16]) -> [f64; 16] {
-    let mut out = [0.0f64; 16];
-    for r in 0..4 {
-        for c in 0..4 {
-            let mut s = 0.0;
-            for k in 0..4 {
-                s += a[r * 4 + k] * b[k * 4 + c];
-            }
-            out[r * 4 + c] = s;
-        }
-    }
-    out
-}
-
-/// Row-major translation matrix.
-fn mat4_translation(t: [f64; 3]) -> [f64; 16] {
-    [
-        1.0, 0.0, 0.0, t[0], //
-        0.0, 1.0, 0.0, t[1], //
-        0.0, 0.0, 1.0, t[2], //
-        0.0, 0.0, 0.0, 1.0,
-    ]
-}
-
-/// Transpose a row-major f64 4x4 into the column-major `[f32; 16]` glTF expects.
-fn row_major_f64_to_col_major_f32(m: &[f64; 16]) -> [f32; 16] {
-    let mut out = [0.0f32; 16];
-    for r in 0..4 {
-        for c in 0..4 {
-            out[c * 4 + r] = m[r * 4 + c] as f32;
-        }
-    }
-    out
-}
-
-/// Inverse of a row-major AFFINE 4x4 (last row `[0,0,0,1]`): invert the upper 3x3
-/// (cofactor / determinant) and map the translation by `-R⁻¹·t`. Returns `None` if
-/// the 3x3 is singular (degenerate placement) so the caller can fall back to flat.
-fn affine_inverse(m: &[f64; 16]) -> Option<[f64; 16]> {
-    let a = m[0]; let b = m[1]; let c = m[2];
-    let d = m[4]; let e = m[5]; let f = m[6];
-    let g = m[8]; let h = m[9]; let i = m[10];
-    let det = a * (e * i - f * h) - b * (d * i - f * g) + c * (d * h - e * g);
-    if det.abs() < 1e-18 {
-        return None;
-    }
-    let inv_det = 1.0 / det;
-    // Inverse of the 3x3 (row-major) via the transposed cofactor matrix.
-    let r = [
-        (e * i - f * h) * inv_det,
-        (c * h - b * i) * inv_det,
-        (b * f - c * e) * inv_det,
-        (f * g - d * i) * inv_det,
-        (a * i - c * g) * inv_det,
-        (c * d - a * f) * inv_det,
-        (d * h - e * g) * inv_det,
-        (b * g - a * h) * inv_det,
-        (a * e - b * d) * inv_det,
-    ];
-    let (tx, ty, tz) = (m[3], m[7], m[11]);
-    // Translation of the inverse: -R⁻¹ · t.
-    let it = [
-        -(r[0] * tx + r[1] * ty + r[2] * tz),
-        -(r[3] * tx + r[4] * ty + r[5] * tz),
-        -(r[6] * tx + r[7] * ty + r[8] * tz),
-    ];
-    Some([
-        r[0], r[1], r[2], it[0], //
-        r[3], r[4], r[5], it[1], //
-        r[6], r[7], r[8], it[2], //
-        0.0, 0.0, 0.0, 1.0,
-    ])
-}
-
-/// Compose an `InstanceMeta`'s world placement `transform · local · canonical`
-/// (row-major f64), the same product the collator's `compose_world` builds.
-fn compose_world_meta(meta: &InstanceMeta) -> [f64; 16] {
-    let local = meta.local_transform.unwrap_or(IDENTITY16);
-    let canonical = meta.canonical_transform.unwrap_or(IDENTITY16);
-    mat4_mul(&meta.transform, &mat4_mul(&local, &canonical))
-}
-
-/// Build the column-major glTF node matrix placing a shared template (Y-up local
-/// geometry, relative to `template_origin_yup`) at one occurrence's BAKED pose.
-/// Recomputed in f64 from the occurrence's `InstanceMeta`, the precomputed template
-/// inverse `m_ref_inv` (`affine_inverse(compose_world_meta(template))`, computed once
-/// per group), and the model `rtc` offset (Z-up) the baker subtracted.
-fn occurrence_node_matrix(
-    occ: &InstanceMeta,
-    m_ref_inv: &[f64; 16],
-    rtc_zup: [f64; 3],
-    template_origin_yup: [f64; 3],
-    scene_center: [f64; 3],
-) -> [f32; 16] {
-    let m_k = compose_world_meta(occ);
-    // rel maps the template's PRE-RTC world geometry onto occurrence k's.
-    let rel_pre = mat4_mul(&m_k, m_ref_inv);
-    // Conjugate into the POST-RTC baked frame the geometry actually lives in.
-    let rel_baked = mat4_mul(
-        &mat4_translation([-rtc_zup[0], -rtc_zup[1], -rtc_zup[2]]),
-        &mat4_mul(&rel_pre, &mat4_translation(rtc_zup)),
-    );
-    // Conjugate Z-up→Y-up (the template was converted by the same S).
-    let rel_yup = mat4_mul(&mat4_mul(&S_YUP, &rel_baked), &S_YUP_INV);
-    let n = mat4_mul(
-        &mat4_translation([-scene_center[0], -scene_center[1], -scene_center[2]]),
-        &mat4_mul(&rel_yup, &mat4_translation(template_origin_yup)),
-    );
-    row_major_f64_to_col_major_f32(&n)
 }
 
 /// Streams geometry into one or more glTF buffers. Each buffer holds three bufferViews
@@ -2600,155 +2456,6 @@ fn write_bounded_glb(
     (out, stats)
 }
 
-/// Fail-closed [`export_glb_from_meshes`]: validates that the per-mesh
-/// `vertex_counts` / `index_counts` (and normals) are fully backed by the flattened
-/// buffers BEFORE assembling. If a count runs past the end of `positions` / `indices`
-/// / `normals`, returns [`ExportError::MalformedMeshInput`] instead of silently
-/// dropping the un-backed tail (which the infallible variant does — a valid GLB
-/// missing part of the model, reported as success). Prefer this at any boundary where
-/// the counts and buffers come from separate sources (the wasm FFI).
-#[allow(clippy::too_many_arguments)]
-pub fn try_export_glb_from_meshes(
-    positions: &[f32],
-    normals: &[f32],
-    indices: &[u32],
-    vertex_counts: &[u32],
-    index_counts: &[u32],
-    colors: &[f32],
-    origins: &[f64],
-    express_ids: &[u32],
-    include_metadata: bool,
-    lit: bool,
-    emissive: bool,
-) -> Result<(Vec<u8>, GltfStats), ExportError> {
-    // Sum the declared counts and check they fit the flattened buffers. `u64` sums so a
-    // caller passing absurd counts can't overflow `usize` on wasm32 before the compare.
-    // `index_counts` is truncated to `vertex_counts.len()`, matching the assembler, which
-    // treats a missing `index_counts[i]` as 0 and ignores any surplus entries.
-    let n = vertex_counts.len();
-    let vsum: u64 = vertex_counts.iter().map(|&c| c as u64).sum();
-    let isum: u64 = index_counts.iter().take(n).map(|&c| c as u64).sum();
-    if vsum * 3 > positions.len() as u64 {
-        return Err(ExportError::MalformedMeshInput {
-            detail: format!(
-                "vertex_counts sum to {vsum} vertices ({} position floats) but `positions` has {}",
-                vsum * 3,
-                positions.len()
-            ),
-        });
-    }
-    // Normals must be either absent (0) or cover every emitted vertex; a short-but-nonzero
-    // normals buffer would make the under-covered meshes silently fail `view_ok` and vanish.
-    if !normals.is_empty() && (normals.len() as u64) < vsum * 3 {
-        return Err(ExportError::MalformedMeshInput {
-            detail: format!(
-                "`normals` has {} floats but the meshes need {} (vertex_counts sum to {vsum})",
-                normals.len(),
-                vsum * 3
-            ),
-        });
-    }
-    if isum > indices.len() as u64 {
-        return Err(ExportError::MalformedMeshInput {
-            detail: format!(
-                "index_counts sum to {isum} but `indices` has {}",
-                indices.len()
-            ),
-        });
-    }
-    // Every count is backed, so the infallible assembler's malformed-input `break` is
-    // unreachable and no mesh is dropped.
-    Ok(export_glb_from_meshes(
-        positions, normals, indices, vertex_counts, index_counts, colors, origins,
-        express_ids, include_metadata, lit, emissive,
-    ))
-}
-
-/// Assemble a GLB from already-produced meshes (the viewer's MeshData — **no re-meshing**).
-/// Per mesh `i`: `vertex_counts[i]` vertices + `index_counts[i]` indices, taken in order
-/// from the concatenated `positions`/`normals`/`indices`; `colors` is RGBA per mesh,
-/// `origins` is xyz per mesh, `express_ids` labels each mesh. Indices are per-mesh LOCAL.
-/// Callers pass exactly the meshes they want emitted (visibility filtering is theirs).
-///
-/// NOTE: on malformed input — a `vertex_counts`/`index_counts` entry that runs past the
-/// end of the flattened buffers — this stops at the first un-backed mesh and returns the
-/// valid prefix, so a caller bug ships a GLB missing the model's tail as "success". Use
-/// [`try_export_glb_from_meshes`] to turn that into [`ExportError::MalformedMeshInput`].
-#[allow(clippy::too_many_arguments)]
-// The index `i` walks several parallel count/offset arrays in lockstep; a
-// range loop is the clearest expression and avoids zipping ragged slices.
-#[allow(clippy::needless_range_loop)]
-pub fn export_glb_from_meshes(
-    positions: &[f32],
-    normals: &[f32],
-    indices: &[u32],
-    vertex_counts: &[u32],
-    index_counts: &[u32],
-    colors: &[f32],
-    origins: &[f64],
-    express_ids: &[u32],
-    include_metadata: bool,
-    lit: bool,
-    emissive: bool,
-) -> (Vec<u8>, GltfStats) {
-    let n = vertex_counts.len();
-    // The viewer's `MeshData` arrives pre-welded from the mesh source
-    // (`ifc_lite_processing::element::build_mesh_data` welds every element via
-    // `ifc_lite_geometry::mesh_weld::weld_indexed`), so this path no longer
-    // re-welds — it slices each mesh's block straight into a borrowing
-    // `MeshView`. Views borrow the caller's buffers, which outlive the call.
-    let mut views: Vec<MeshView> = Vec::with_capacity(n);
-    let mut vbase = 0usize; // running vertex offset
-    let mut ibase = 0usize; // running index offset
-    for i in 0..n {
-        let vc = vertex_counts[i] as usize;
-        let ic = index_counts.get(i).copied().unwrap_or(0) as usize;
-        if (vbase + vc) * 3 > positions.len() || ibase + ic > indices.len() {
-            break; // malformed counts — stop rather than panic
-        }
-        let pslice = &positions[vbase * 3..(vbase + vc) * 3];
-        let nslice: &[f32] = if normals.len() >= (vbase + vc) * 3 {
-            &normals[vbase * 3..(vbase + vc) * 3]
-        } else {
-            &[]
-        };
-        let islice = &indices[ibase..ibase + ic];
-        let color = [
-            colors.get(i * 4).copied().unwrap_or(0.8),
-            colors.get(i * 4 + 1).copied().unwrap_or(0.8),
-            colors.get(i * 4 + 2).copied().unwrap_or(0.8),
-            colors.get(i * 4 + 3).copied().unwrap_or(1.0),
-        ];
-        let origin = [
-            origins.get(i * 3).copied().unwrap_or(0.0),
-            origins.get(i * 3 + 1).copied().unwrap_or(0.0),
-            origins.get(i * 3 + 2).copied().unwrap_or(0.0),
-        ];
-        views.push(MeshView {
-            express_id: express_ids.get(i).copied().unwrap_or(0),
-            ifc_type: "",
-            global_id: None,
-            positions: pslice,
-            normals: nslice,
-            indices: islice,
-            color,
-            origin,
-            // The viewer's MeshData drops the instancing side-channel across the
-            // worker boundary (it is `#[serde(skip)]`), so this path is always flat.
-            instance: None,
-        });
-        vbase += vc;
-        ibase += ic;
-    }
-    // From-meshes geometry is already absolute Y-up and never instances (no
-    // side-channel), so there is no RTC frame to compensate. Quantization is a
-    // from-bytes feature; the viewer path stays f32.
-    let mut ch = Chunker::new(12, usize::MAX, None);
-    let (gltf, stats) =
-        build_gltf(&views, include_metadata, None, lit, emissive, [0.0, 0.0, 0.0], false, &mut ch);
-    let json = serde_json::to_vec(&gltf).expect("glTF JSON serializes");
-    (pack_glb(&json, &ch.pos, &ch.norm, &ch.idx), stats)
-}
 
 /// Pack a glTF JSON document and the binary buffer's three runs (positions | normals |
 /// indices) into a GLB container (little-endian). The runs are appended straight into the
@@ -3594,6 +3301,49 @@ mod tests {
     }
 
     #[test]
+    fn try_from_meshes_rejects_short_normals() {
+        // One quad (4 verts) needs 12 normal floats, but only 6 are supplied — the
+        // under-covered mesh would fall back to an empty normal slice and vanish.
+        let positions: Vec<f32> = vec![0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 1.0, 0.0, 0.0, 1.0, 0.0];
+        let normals: Vec<f32> = vec![0.0, 0.0, 1.0, 0.0, 0.0, 1.0]; // short of vsum*3 = 12
+        let indices: Vec<u32> = vec![0, 1, 2, 0, 2, 3];
+        let (vc, ic) = (vec![4u32], vec![6u32]);
+        let color = vec![0.5, 0.5, 0.5, 1.0];
+        let origin = vec![0.0, 0.0, 0.0];
+        let ids = vec![1u32];
+        let err = try_export_glb_from_meshes(
+            &positions, &normals, &indices, &vc, &ic, &color, &origin, &ids, false, true, false,
+        )
+        .expect_err("short normals must be MalformedMeshInput");
+        assert!(matches!(err, ExportError::MalformedMeshInput { .. }), "got {err:?}");
+
+        // Empty normals is the degenerate case that would drop the WHOLE model.
+        let err_empty = try_export_glb_from_meshes(
+            &positions, &[], &indices, &vc, &ic, &color, &origin, &ids, false, true, false,
+        )
+        .expect_err("empty normals must be MalformedMeshInput");
+        assert!(matches!(err_empty, ExportError::MalformedMeshInput { .. }), "got {err_empty:?}");
+    }
+
+    #[test]
+    fn try_from_meshes_rejects_index_counts_past_buffer() {
+        // `index_counts` declares more indices than `indices` holds — the overrun the
+        // infallible path's per-mesh `break` would silently truncate at.
+        let positions: Vec<f32> = vec![0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 1.0, 0.0, 0.0, 1.0, 0.0];
+        let normals: Vec<f32> = vec![0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0];
+        let indices: Vec<u32> = vec![0, 1, 2, 0, 2, 3];
+        let (vc, ic) = (vec![4u32], vec![12u32]); // 12 > indices.len() == 6
+        let color = vec![0.5, 0.5, 0.5, 1.0];
+        let origin = vec![0.0, 0.0, 0.0];
+        let ids = vec![1u32];
+        let err = try_export_glb_from_meshes(
+            &positions, &normals, &indices, &vc, &ic, &color, &origin, &ids, false, true, false,
+        )
+        .expect_err("index_counts past `indices` must be MalformedMeshInput");
+        assert!(matches!(err, ExportError::MalformedMeshInput { .. }), "got {err:?}");
+    }
+
+    #[test]
     fn export_is_byte_deterministic() {
         // Instancing groups by HashMap keys (rep colour buckets, material dedup);
         // emission order must be fixed so repeated exports are byte-identical.
@@ -3706,122 +3456,6 @@ mod tests {
                  (run `pnpm fixtures`)"
             );
         }
-    }
-
-    #[test]
-    fn occurrence_matrix_reconstructs_rotated_instance_under_national_grid_rtc() {
-        // Decisive synthetic test for the RTC/rotation frame (review finding C1+M1):
-        // a ROTATED occurrence at NATIONAL-GRID coordinates. The node matrix is built
-        // from the same InstanceMeta the baker would carry; reconstructing the
-        // occurrence from the template's baked-local geometry must land on the
-        // occurrence's own baked geometry to sub-millimetre, even though the relative
-        // transform's absolute terms are ~1e5 m. (A pre-RTC `rel` applied to post-RTC
-        // geometry — the bug — misplaces this by ~(R-I)·rtc, i.e. hundreds of metres.)
-        use ifc_lite_geometry::InstanceMeta;
-
-        // Row-major helpers.
-        let translate = |t: [f64; 3]| -> [f64; 16] {
-            [1., 0., 0., t[0], 0., 1., 0., t[1], 0., 0., 1., t[2], 0., 0., 0., 1.]
-        };
-        // Rotation about Z (Z-up): (x,y) rotate, z fixed.
-        let rot_z = |deg: f64| -> [f64; 16] {
-            let (s, c) = (deg.to_radians().sin(), deg.to_radians().cos());
-            [c, -s, 0., 0., s, c, 0., 0., 0., 0., 1., 0., 0., 0., 0., 1.]
-        };
-        let apply = |m: &[f64; 16], p: [f64; 3]| -> [f64; 3] {
-            [
-                m[0] * p[0] + m[1] * p[1] + m[2] * p[2] + m[3],
-                m[4] * p[0] + m[5] * p[1] + m[6] * p[2] + m[7],
-                m[8] * p[0] + m[9] * p[1] + m[10] * p[2] + m[11],
-            ]
-        };
-
-        // Placements (Z-up, pre-RTC): template upright, occurrence rotated 37° about Z.
-        let m_ref = super::mat4_mul(&translate([10., 20., 5.]), &rot_z(0.0));
-        let m_k = super::mat4_mul(&translate([60., 35., 5.]), &rot_z(37.0));
-        // National-grid RTC the baker subtracts (e.g. Dutch RD-ish easting/northing).
-        let rtc = [155_000.0_f64, 463_000.0, 0.0];
-
-        // Canonical (rep-local) geometry — a few non-degenerate points.
-        let canonical = [
-            [0.0, 0.0, 0.0],
-            [1.0, 0.0, 0.0],
-            [0.0, 2.0, 0.0],
-            [0.5, 0.5, 1.5],
-            [2.0, 0.3, 0.7],
-        ];
-        // Baked = placement·canonical - rtc, in Z-up.
-        let bake = |m: &[f64; 16]| -> Vec<[f64; 3]> {
-            canonical
-                .iter()
-                .map(|&x| {
-                    let w = apply(m, x);
-                    [w[0] - rtc[0], w[1] - rtc[1], w[2] - rtc[2]]
-                })
-                .collect()
-        };
-        let tmpl_baked = bake(&m_ref);
-        let occ_baked = bake(&m_k);
-
-        // Template origin = centroid of its baked geometry; local = baked - origin.
-        let n = canonical.len() as f64;
-        let origin_z = {
-            let mut o = [0.0; 3];
-            for p in &tmpl_baked {
-                for k in 0..3 {
-                    o[k] += p[k] / n;
-                }
-            }
-            o
-        };
-        // Convert template origin + local, and the occurrence's baked truth, to Y-up.
-        let origin_yup = crate::frame::yup_f64(origin_z);
-        let tmpl_local_yup: Vec<[f64; 3]> = tmpl_baked
-            .iter()
-            .map(|p| crate::frame::yup_f64([p[0] - origin_z[0], p[1] - origin_z[1], p[2] - origin_z[2]]))
-            .collect();
-        let occ_world_yup: Vec<[f64; 3]> = occ_baked.iter().map(|p| crate::frame::yup_f64(*p)).collect();
-
-        // scene_center = centre of the combined baked Y-up AABB.
-        let mut lo = [f64::INFINITY; 3];
-        let mut hi = [f64::NEG_INFINITY; 3];
-        for p in tmpl_baked.iter().chain(occ_baked.iter()) {
-            let y = crate::frame::yup_f64(*p);
-            for k in 0..3 {
-                lo[k] = lo[k].min(y[k]);
-                hi[k] = hi[k].max(y[k]);
-            }
-        }
-        let scene_center = [(lo[0] + hi[0]) * 0.5, (lo[1] + hi[1]) * 0.5, (lo[2] + hi[2]) * 0.5];
-
-        let meta = |transform: [f64; 16]| InstanceMeta {
-            transform,
-            local_transform: None,
-            canonical_transform: None,
-            rep_identity: 42,
-            instanceable: true,
-        };
-        let m_ref_inv = super::affine_inverse(&super::compose_world_meta(&meta(m_ref)))
-            .expect("template placement invertible");
-        let node = super::occurrence_node_matrix(&meta(m_k), &m_ref_inv, rtc, origin_yup, scene_center);
-
-        // Reconstruct: world = scene_center(root) + node(col-major) · template_local.
-        let mut max_err = 0.0f64;
-        for (lv, truth) in tmpl_local_yup.iter().zip(&occ_world_yup) {
-            let (x, y, z) = (lv[0], lv[1], lv[2]);
-            let world = [
-                scene_center[0] + node[0] as f64 * x + node[4] as f64 * y + node[8] as f64 * z + node[12] as f64,
-                scene_center[1] + node[1] as f64 * x + node[5] as f64 * y + node[9] as f64 * z + node[13] as f64,
-                scene_center[2] + node[2] as f64 * x + node[6] as f64 * y + node[10] as f64 * z + node[14] as f64,
-            ];
-            for k in 0..3 {
-                max_err = max_err.max((world[k] - truth[k]).abs());
-            }
-        }
-        assert!(
-            max_err < 1e-3,
-            "rotated instance under national-grid RTC mis-reconstructed by {max_err} m"
-        );
     }
 
     #[test]
